Create weibull
File for the Weibull distribution and the wind speed as a function of height. It requires the use of some text files that will be uploaded to drive in the "background" folder. You can make a Weibull fit for anything: average wind speed, maximum wind speed, temperature, direction... simply change the v=... before the weibull fit section